Fix minor bugs and CodeFactor Added manual DI injection
