+++ OpenTabletDriver.Desktop/Reflection/IServiceManager.cs
@@ -6,5 +6,6 @@ namespace OpenTabletDriver.Desktop.Reflection
     {
         bool AddService<T>(Func<T> value);
         void ResetServices();
+        void Inject(object obj);
     }
 }
